# resources/swagger/swagger.yml
info:
  title: Comentario
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
swagger: "2.0"
basePath: /api
securityDefinitions:
  # Header authentication for commenters (used by the embed part)
  userSessionHeader:
    type: apiKey
    in: header
    name: X-User-Session

  # Cookie authentication for users (used by the frontend part). Uses the apiKey type for the lack of a proper
  # cookie-based authentication in Swagger 2
  userCookie:
    type: apiKey
    in: header
    name: Cookie

  # Token authentication for certain endpoints, called externally, such as password reset or email confirmation
  token:
    type: oauth2
    flow: accessCode
    authorizationUrl: http://dummy/
    tokenUrl: http://dummy/
    scopes:
      confirm-email: confirm user's email
      login: authenticate the user
      pwd-reset: reset user's password

# Default security is cookie-based user authentication
security:
  - userCookie: []
...
parameters:
  federatedIdpId:
    in: path
    name: provider
    required: true
    description: Federated identity provider ID. The same as the federatedIdpId type, but also includes 'sso'
    type: string
    maxLength: 37
    pattern: '^facebook|github|gitlab|google|twitter|sso|(oidc:[-a-z0-9]{1,32})$'
...
  queryOptionalHost:
    name: host
    in: query
    required: false
    description: Optional host query parameter
    type: string
    minLength: 1
    maxLength: 259
    pattern: "[-.a-z0-9]{1,253}(:[0-9]{1-5})?"
...
  queryToken:
    in: query
    name: token
    required: true
    description: ID consisting of 64 hex digits, in the query parameter
    type: string
    minLength: 64
    maxLength: 64
    pattern: '[0-9a-f]{64}'
...
responses:
...
  OAuthRedirect:
    description: Redirecting to a federated identity provider
    headers:
      Location:
        type: string

  # 401
  OAuthFailure:
    description: Authentication failed
    schema:
      type: string
paths:
...
  /oauth/{provider}:
    get:
      operationId: AuthOauthInit
      summary: Initiate signing in via federated authentication using an anonymous token
      tags:
        - ApiGeneral
      security: []
      produces:
        - text/html # For an error response only, otherwise a redirect is issued
      parameters:
        - $ref: "#/parameters/federatedIdpId"
        - $ref: "#/parameters/queryOptionalHost"
        - $ref: "#/parameters/queryToken"
      responses:
        307:
          $ref: "#/responses/OAuthRedirect"
        401:
          $ref: "#/responses/OAuthFailure"
  /oauth/{provider}/callback:
    get:
      operationId: AuthOauthCallback
      summary: Endpoint that gets called back by the federated authentication provider
      tags:
        - ApiGeneral
      security: []
      produces:
        - text/html
      parameters:
        - $ref: "#/parameters/federatedIdpId"
      responses:
        200:
          description: User has successfully signed in
        401:
          $ref: "#/responses/OAuthFailure"

  #---------------------------------------------------------------------------------------------------------------------
  # Current user
  #---------------------------------------------------------------------------------------------------------------------
...